# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Klaro - AI-powered business management platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface SignupBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export type RawInputProcessingStatus =
  (typeof RawInputProcessingStatus)[keyof typeof RawInputProcessingStatus];

export const RawInputProcessingStatus = {
  pending: "pending",
  processing: "processing",
  done: "done",
  failed: "failed",
} as const;

export interface RawInput {
  id: number;
  userId: number;
  fileName: string;
  fileType: string;
  filePath: string;
  processingStatus: RawInputProcessingStatus;
  parsedRecordCount: number;
  createdAt: string;
}

export type RawInputDetailProcessingStatus =
  (typeof RawInputDetailProcessingStatus)[keyof typeof RawInputDetailProcessingStatus];

export const RawInputDetailProcessingStatus = {
  pending: "pending",
  processing: "processing",
  done: "done",
  failed: "failed",
} as const;

export type ParsedRecordType =
  (typeof ParsedRecordType)[keyof typeof ParsedRecordType];

export const ParsedRecordType = {
  income: "income",
  expense: "expense",
} as const;

export interface ParsedRecord {
  id: number;
  rawInputId: number;
  userId: number;
  date: string;
  description: string;
  amount: number;
  type: ParsedRecordType;
  category: string;
  /** @nullable */
  quantity?: number | null;
  /** @nullable */
  confidence?: number | null;
  isConfirmed: boolean;
  createdAt: string;
}

export interface RawInputDetail {
  id: number;
  userId: number;
  fileName: string;
  fileType: string;
  filePath: string;
  processingStatus: RawInputDetailProcessingStatus;
  parsedRecords: ParsedRecord[];
  createdAt: string;
}

export type CreateParsedRecordBodyType =
  (typeof CreateParsedRecordBodyType)[keyof typeof CreateParsedRecordBodyType];

export const CreateParsedRecordBodyType = {
  income: "income",
  expense: "expense",
} as const;

export interface CreateParsedRecordBody {
  date: string;
  description: string;
  amount: number;
  type: CreateParsedRecordBodyType;
  category: string;
  /** @nullable */
  quantity?: number | null;
}

export type UpdateParsedRecordBodyType =
  (typeof UpdateParsedRecordBodyType)[keyof typeof UpdateParsedRecordBodyType];

export const UpdateParsedRecordBodyType = {
  income: "income",
  expense: "expense",
} as const;

export interface UpdateParsedRecordBody {
  date?: string;
  description?: string;
  amount?: number;
  type?: UpdateParsedRecordBodyType;
  category?: string;
  /** @nullable */
  quantity?: number | null;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  income: "income",
  expense: "expense",
} as const;

export interface Transaction {
  id: number;
  userId: number;
  date: string;
  description: string;
  amount: number;
  type: TransactionType;
  category: string;
  /** @nullable */
  quantity?: number | null;
  /** @nullable */
  sourceRawInputId?: number | null;
  createdAt: string;
}

export interface ConfirmResult {
  confirmedCount: number;
  transactions: Transaction[];
}

export interface Insight {
  id: number;
  userId: number;
  title: string;
  description: string;
  recommendation: string;
  periodLabel: string;
  createdAt: string;
}

export interface DashboardSummary {
  totalIncome: number;
  totalExpenses: number;
  netBalance: number;
  transactionCount: number;
  uploadCount: number;
  /** @nullable */
  topCategory?: string | null;
}

export interface MonthlyTrendPoint {
  month: string;
  income: number;
  expenses: number;
}

export interface CategoryBreakdown {
  category: string;
  total: number;
  count: number;
}

export type UploadFileBody = {
  file: Blob;
};

export type ListParsedRecordsParams = {
  /**
   * @nullable
   */
  rawInputId?: number | null;
  /**
   * @nullable
   */
  confirmed?: string | null;
};

export type CreateParsedRecordsBody = {
  records: CreateParsedRecordBody[];
  rawInputId: number;
};

export type ConfirmParsedRecordsBody = {
  rawInputId: number;
};

export type ListTransactionsParams = {
  /**
   * @nullable
   */
  limit?: number | null;
  /**
   * @nullable
   */
  offset?: number | null;
  /**
   * @nullable
   */
  category?: string | null;
  /**
   * @nullable
   */
  type?: string | null;
};