# lib/api-zod/src/generated/types/parsedRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Klaro - AI-powered business management platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ParsedRecordType } from "./parsedRecordType";

export interface ParsedRecord {
  id: number;
  rawInputId: number;
  userId: number;
  date: string;
  description: string;
  amount: number;
  type: ParsedRecordType;
  category: string;
  /** @nullable */
  quantity?: number | null;
  /** @nullable */
  confidence?: number | null;
  isConfirmed: boolean;
  createdAt: Date;
}